+++ CHANGELOG.md
@@ -24,6 +24,7 @@ CHANGELOG for FlatCAM beta
 - more refactoring; solved some issues introduced by the refactoring
 - solved a circular import
 - updated the language translation files to the latest changes (no translation)
+- working on a new Tool: Etch Compensation Tool -> installed the tool and created the GUI and class template
 
 17.05.2020